+++ app/controllers/frontend/dnd2024/spells_controller.rb
@@ -5,13 +5,17 @@ module Frontend
     class SpellsController < Frontend::BaseController
       include SerializeRelation
 
-      INDEX_SERIALIZER_FIELDS = %i[id slug name level available_for].freeze
-
       def index
-        serialize_relation(relation, ::Dnd2024::SpellSerializer, :spells, only: INDEX_SERIALIZER_FIELDS)
+        serialize_relation_v2(
+          relation, ::Dnd2024::SpellSerializer, :spells, cache_options: cache_options, order_options: { key: 'name' }
+        )
       end
 
       private
+
+      def cache_options
+        { key: ["spells/dnd2024/#{I18n.locale}/v1", params[:max_level]].compact.join('/'), expires_in: 12.hours }
+      end
 
       def relation
         relation = ::Spell.dnd2024