+++ app/controllers/frontend/dnd5/spells_controller.rb
@@ -8,10 +8,16 @@ module Frontend
       INDEX_SERIALIZER_FIELDS = %i[id slug name level available_for].freeze
 
       def index
-        serialize_relation(relation, ::Dnd5::SpellSerializer, :spells, only: INDEX_SERIALIZER_FIELDS)
+        serialize_relation_v2(
+          relation, ::Dnd5::SpellSerializer, :spells, cache_options: cache_options, order_options: { key: 'name' }
+        )
       end
 
       private
+
+      def cache_options
+        { key: ["spells/dnd5/#{I18n.locale}/v1", params[:max_level]].compact.join('/'), expires_in: 12.hours }
+      end
 
       def relation
         relation = ::Spell.dnd5